+++ src/bindings/bnd_xform.cpp
@@ -33,6 +33,13 @@ BND_Transform BND_Transform::Rotation(double angleRadians, ON_3dVector rotationA
   return rc;
 }
 
+BND_Transform BND_Transform::RotationFromTwoVectors(ON_3dVector startDirection, ON_3dVector endDirection, ON_3dPoint rotationCenter)
+{
+  BND_Transform rc(1);
+  rc.m_xform.Rotation(startDirection, endDirection, rotationCenter);
+  return rc;
+}
+
 BND_Transform BND_Transform::Mirror(ON_3dPoint pointOnMirrorPlane, ON_3dVector normalToMirrorPlane)
 {
   BND_Transform rc(1);
@@ -45,6 +52,18 @@ BND_Transform BND_Transform::Mirror2(BND_Plane mirrorPlane)
   BND_Transform rc(1);
   rc.m_xform.Mirror(pl.Origin(), pl.zaxis);
   return rc;
+}
+
+BND_Transform BND_Transform::PlaneToPlane(BND_Plane plane0, BND_Plane plane1)
+{
+  BND_Transform rc(1);
+  rc.m_xform.Rotation(plane0.ToOnPlane(), plane1.ToOnPlane());
+  return rc;
+}
+
+BND_Transform BND_Transform::Shear(BND_Plane plane, ON_3dVector x, ON_3dVector y, ON_3dVector z)
+{
+  return BND_Transform(ON_Xform::ShearTransformation(plane.ToOnPlane(), x, y, z));
 }
 
 BND_Transform BND_Transform::Multiply(BND_Transform a, BND_Transform b)
@@ -110,14 +129,22 @@ void initXformBindings(pybind11::module& m)
     .def_static("Scale", &BND_Transform::Scale, py::arg("anchor"), py::arg("scaleFactor"))
     .def_static("Scale", &BND_Transform::Scale2, py::arg("plane"), py::arg("xScaleFactor"), py::arg("yScaleFactor"), py::arg("zScaleFactor"))
     .def_static("Rotation", &BND_Transform::Rotation, py::arg("angleRadians"), py::arg("rotationAxis"), py::arg("rotationCenter"))
+    .def_static("Rotation", &BND_Transform::RotationFromTwoVectors, py::arg("startDirection"), py::arg("endDirection"), py::arg("rotationCenter"))
     .def_static("Mirror", &BND_Transform::Mirror, py::arg("pointOnMirrorPlane"), py::arg("normalToMirrorPlane"))
     .def_static("Mirror", &BND_Transform::Mirror2, py::arg("mirrorPlane"))
+    .def_static("PlaneToPlane", &BND_Transform::PlaneToPlane, py::arg("plane0"), py::arg("plane1"))
+    .def_static("Shear", &BND_Transform::Shear, py::arg("plane"), py::arg("x"), py::arg("y"), py::arg("z"))
     .def_static("Multiply", &BND_Transform::Multiply, py::arg("a"), py::arg("b"))
+    .def_property_readonly("IsAffine", &BND_Transform::IsAffine)
     .def_property_readonly("IsIdentity", &BND_Transform::IsIdentity)
+    .def_property_readonly("IsLinear", &BND_Transform::IsLinear)
+    .def_property_readonly("IsRotation", &BND_Transform::IsRotation)
     .def_property_readonly("IsValid", &BND_Transform::IsValid)
     .def_property_readonly("IsZero", &BND_Transform::IsZero)
     .def_property_readonly("IsZero4x4", &BND_Transform::IsZero4x4)
     .def_property_readonly("IsZeroTransformation", &BND_Transform::IsZeroTransformation)
+    .def_property_readonly("RigidType", &BND_Transform::RigidType)
+    .def_property_readonly("SimilarityType", &BND_Transform::SimilarityType)
     .def("Determinant", &BND_Transform::Determinant)
     .def("TryGetInverse", &BND_Transform::TryGetInverse)
     .def("TransformBoundingBox", &BND_Transform::TransformBoundingBox, py::arg("bbox"))
@@ -140,6 +167,18 @@ void initXformBindings(pybind11::module& m)
     .def_property("M32", &BND_Transform::GetM32, &BND_Transform::SetM32)
     .def_property("M33", &BND_Transform::GetM33, &BND_Transform::SetM33)
     ;
+
+  py::enum_<TransformSimilarityType>(m, "TransformSimilarityType")
+    .value("OrientationReversing", TransformSimilarityType::OrientationReversing)
+    .value("NotSimilarity", TransformSimilarityType::NotSimilarity)
+    .value("OrientationPreserving", TransformSimilarityType::OrientationPreserving)
+    ;
+
+  py::enum_<TransformRigidType>(m, "TransformRigidType")
+    .value("RigidReversing", TransformRigidType::RigidReversing)
+    .value("NotRigid", TransformRigidType::NotRigid)
+    .value("Rigid", TransformRigidType::Rigid)
+    ;
 }
 #endif
 
@@ -148,6 +187,19 @@ using namespace emscripten;
 
 void initXformBindings(void*)
 {
+
+  enum_<TransformSimilarityType>("TransformSimilarityType")
+    .value("OrientationReversing", TransformSimilarityType::OrientationReversing)
+    .value("NotSimilarity", TransformSimilarityType::NotSimilarity)
+    .value("OrientationPreserving", TransformSimilarityType::OrientationPreserving)
+    ;
+
+  enum_<TransformRigidType>("TransformRigidType")
+    .value("RigidReversing", TransformRigidType::RigidReversing)
+    .value("NotRigid", TransformRigidType::NotRigid)
+    .value("Rigid", TransformRigidType::Rigid)
+    ;
+
   class_<BND_Transform>("Transform")
     .constructor<double>()
     //.constructor<const BND_Transform&>()
@@ -155,14 +207,25 @@ void initXformBindings(void*)
     .class_function("zeroTransformation", &BND_Transform::ZeroTransformation)
     .class_function("unset", &BND_Transform::Unset)
     .class_function("translation", &BND_Transform::Translation)
+    .class_function("translation", &BND_Transform::Translation1)
     .class_function("scale", &BND_Transform::Scale)
     .class_function("rotation", &BND_Transform::Rotation)
+    .class_function("rotation", &BND_Transform::RotationFromTwoVectors)
+    .class_function("mirror", &BND_Transform::Mirror)
+    .class_function("mirror", &BND_Transform::Mirror2)
+    .class_function("planeToPlane", &BND_Transform::PlaneToPlane)
+    .class_function("shear", &BND_Transform::Shear)
     .class_function("multiply", &BND_Transform::Multiply)
+    .property("isAffine", &BND_Transform::IsAffine)
     .property("isIdentity", &BND_Transform::IsIdentity)
+    .property("isLinear", &BND_Transform::IsLinear)
+    .property("isRotation", &BND_Transform::IsRotation)
     .property("isValid", &BND_Transform::IsValid)
     .property("isZero", &BND_Transform::IsZero)
     .property("isZero4x4", &BND_Transform::IsZero4x4)
     .property("isZeroTransformation", &BND_Transform::IsZeroTransformation)
+    .property("rigidType", &BND_Transform::RigidType)
+    .property("similarityType", &BND_Transform::SimilarityType)
     .function("determinant", &BND_Transform::Determinant)
     .function("tryGetInverse", &BND_Transform::TryGetInverse, allow_raw_pointers())
     .function("transformBoundingBox", &BND_Transform::TransformBoundingBox, allow_raw_pointers())